# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Athena management API",
    "version": "1.5.0",
    "description": "Private knowledge base. Create a scoped key in /admin/athena/apikey/. Keys expire after 90 days by default and are revoked when the owning user is disabled, removed, or has their password reset. Browser cookies are not accepted by the API. Guide for agents: GET /api/docs/. Check a key with GET /me/."
  },
  "servers": [
    {
      "url": "https://athena.moratechnology.com/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/articles/": {
      "get": {
        "operationId": "list_articles",
        "summary": "List articles",
        "responses": {
          "200": {
            "description": "JSON {count, offset, limit, results}. Default limit 50, maximum 100."
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid, expired, or revoked key"
          },
          "403": {
            "description": "Insufficient scope or user permissions"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "412": {
            "description": "Missing or stale If-Match"
          }
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "post": {
        "operationId": "create_articles",
        "summary": "Create article",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Invalid, expired, or revoked key"
          },
          "403": {
            "description": "Insufficient scope or user permissions"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Conflict"
          },
          "412": {
            "description": "Missing or stale If-Match"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ArticleInput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "body": {
            "type": "string",
            "description": "Markdown, at most 1 MiB. Required unless the article already has a PDF."
          },
          "status": {
            "type": "string",
            "enum": [
              "",
              "stable",
              "alpha",
              "coming-soon"
            ]
          },
          "url": {
            "type": "string"
          },
          "download_file": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "page",
              "guide",
              "tool",
              "announcement",
              "release"
            ]
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "published": {
            "type": "boolean",
            "default": false
          },
          "pdf_only": {
            "type": "boolean",
            "default": false,
            "description": "The attached PDF is the whole article; the reader offers no Exportar PDF button."
          },
          "is_public": {
            "type": "boolean",
            "default": false,
            "description": "Allow anonymous reading of the published article, its images, attached PDFs, and generated PDF exports. Drafts stay restricted. Existing API authentication and edit permissions still apply."
          }
        }
      },
...
    }
  }
}